+++ src/utilities.jl
@@ -52,16 +52,21 @@ function cor(x::Matrix{Float64})
     return cormat
 end
 
-function euclidean(v1::Array{T1,1}, v2::Array{T2,1})::Float64 where {T1 <: Number,T2 <: Number} 
+function euclidean(v1::Vector, v2::Vector)
     (v1 .- v2).^2.0 |> sum |> sqrt
 end
 
-function euclidean(v1::Array{T,1})::Float64 where T <: Number
+function euclidean(v1::Vector)
     v2 = zeros(Float64, length(v1))
     return euclidean(v1, v2)
 end
 
-function normalize(v1::Array{T,1})::Array{Float64,1} where T <: Number
+function euclidean(v1::Vector, row::DataFrames.DataFrameRow{DataFrame, DataFrames.Index})
+    v2 = Vector(row)
+    (v1 .- v2).^2.0 |> sum |> sqrt
+end
+
+function normalize(v1::Vector)::Vector 
     return v1 ./ euclidean(v1) 
 end
 
@@ -75,12 +80,12 @@ function normalize(data::DataFrame)::DataFrame
 end
 
 
-function apply_columns(fs::Array{Function,1}, data::Union{DataFrame,Array{T,2}} where T <: Number)
+function apply_columns(fs::Array{Function,1}, data)
     _, m = size(data)
     return [fs[i](data[:,i]) for i in 1:m]
 end
 
-function apply_columns(f::Function, data::Union{DataFrame,Array{T,2}} where T <: Number)
+function apply_columns(f::Function, data)
     return [f(c) for c in eachcol(data)]
 end
 
@@ -100,58 +105,68 @@ function colmeans(data)::Array{Number,1}
     return apply_columns(mean, data)
 end
 
-function apply_rows(f::Function, data::Union{DataFrame,Array{T,2}} where T <: Number)
+function apply_rows(f::Function, data)
     return [f(c) for c in eachrow(data)]
 end
 
-function rowmins(data::DataFrame)::Array{Float64,1}
+function rowmins(data::DataFrame)::Vector
     return apply_rows(minimum, data)
 end
 
-function rowmaxs(data::DataFrame)::Array{Float64,1}
+function rowmaxs(data::DataFrame)::Vector
     return apply_rows(maximum, data)
 end
 
-function rowsums(data)::Array{Float64,1}
+function rowsums(data)::Vector
     return apply_rows(sum, data)
 end
 
-function rowmeans(data)::Array{Float64,1}
+function rowmeans(data)::Vector
     return apply_rows(mean, data)
 end
 
-function unitize(v::Array{T,1})::Array{Float64,1} where {T <: Number}
+function unitize(v::Vector)::Vector 
     return v ./ sum(v)
 end
 
-function Base.:*(w::Array{Float64,1}, data::DataFrame)::DataFrame 
+function Base.:*(w::Vector, data::DataFrame)::DataFrame 
     newdf = copy(data)
     _, p = size(newdf)
     for i in 1:p
-        newdf[!, i] = convert.(Float64, newdf[!,i])
+        newdf[!, i] = newdf[!,i]
         newdf[:, i] = w[i] .* data[:, i]
     end
     return newdf
 end
 
-function Base.:-(r1::DataFrameRow, r2::DataFrameRow)::Array{Number,1}
+function Base.:-(r1::DataFrameRow, r2::DataFrameRow)::Vector
     # v1 = convert(Array{Float64,1}, r1)
-    v1 = Vector{Float64}(r1)
+    v1 = Vector(r1)
     # v2 = convert(Array{Float64,1}, r2)
-    v2 = Vector{Float64}(r2)
+    v2 = Vector(r2)
     return v1 .- v2
 end
 
-function Base.:-(r1::Array{T,1}, r2::DataFrameRow)::Array{T,1} where T <: Number
+#function Base.:-(r1::Array{T,1}, r2::DataFrameRow)::Array{T,1} where T <: Number
+#    # v2 = convert(Array{Float64,1}, r2)
+#    v2 = Vector{Float64}(r2)
+#    return r1 .- v2
+#end
+#
+#function Base.:-(r1::DataFrameRow, r2::Array{T,1})::Array{T,1} where T <: Number
+#    v1 = convert(Array{Float64,1}, r1)
+#    return v1 .- r2
+#end
+
+function Base.:-(r1::Vector, r2::DataFrameRow)::Vector
     # v2 = convert(Array{Float64,1}, r2)
-    v2 = Vector{Float64}(r2)
-    return r1 .- v2
+    return r1 .- r2
 end
 
-function Base.:-(r1::DataFrameRow, r2::Array{T,1})::Array{T,1} where T <: Number
-    v1 = convert(Array{Float64,1}, r1)
-    return v1 .- r2
+function Base.:-(r1::DataFrameRow, r2::Vector)::Vector
+    return r1 .- r2
 end
+
 
 """
     makeDecisionMatrix(mat; names)
@@ -163,7 +178,7 @@ end
  - `names::Union{Nothing,Array{String,1}}`: Column names. Default is nothing. If column names are not given, they are labelled as Crt 1, Crt 2, ..., etc.
  
 """
-function makeDecisionMatrix(mat::Array{T,2}; names::Union{Nothing,Array{String,1}}=nothing)::DataFrame where {T <: Number}
+function makeDecisionMatrix(mat::Matrix; names::Union{Nothing,Array{String,1}}=nothing)::DataFrame
     _, m = size(mat)
     df = DataFrame()
     for i in 1:m
@@ -172,7 +187,7 @@ function makeDecisionMatrix(mat::Array{T,2}; names::Union{Nothing,Array{String,1
         else
             name = names[i]
         end
-        df[:,Symbol(name)] = convert(Array{Float64,1}, mat[:, i])     
+        df[:,Symbol(name)] = mat[:, i]
     end
     return df
 end